arch/x86,cpu/x86/mp_init: Switch cpu_info index type to size_t
The alignment for `struct cpu_info` is wrong on x86_64. c_start.S uses
the `push` instruction when setting up the cpu_info struct. This
instruction will push 8 bytes but `unsigned int` is 4 bytes. By making
it a `size_t` we get the correct size for both x86_32 and x86_64.

BUG=b:179699789
TEST=Boot guybrush to the OS

Suggested-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/56573
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/arch/x86/include/arch/cpu.h
@@ -225,7 +225,7 @@ struct thread;
 
 struct cpu_info {
 	struct device *cpu;
-	unsigned int index;
+	size_t index;
 #if CONFIG(COOP_MULTITASKING)
 	struct thread *thread;
 #endif
